move remaining modern components to other
