+++ src/components/legacy/Tabs/Tabs.js
@@ -77,7 +77,7 @@ const TabComponent = ({ name, onPress, active }: TabProps) => {
 /**
  * @deprecated This compontent is considered legacy and should not be used in new code
  *
- * Use: components/modern/Tabs or components/layout/TabBar instead
+ * Use: components/other/Tabs or components/layout/TabBar instead
  */
 const Tabs = ({ tabs, activeTab, wrapperStyle }: Props) => {
   return (
